+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added starknet support for Xverse wallet. 
 - Redirect link to mempool transaction hash when deposit is not confirmed.
 
 ### Fixed
@@ -11,6 +12,7 @@
 - Update toast for successful swap and add abort controller to fetch assets function.
 - Update the coinbase wallet id.
 - Fixed the link redirecting to staking docs.
+- Disable Core chain in output-asset on Phantom, Keplr and Leap wallet.
 - Update title to "Swap completed" from "Swap in progress" when swap is completed.
 
 ## [1.1.7]
@@ -33,6 +35,7 @@
 - Updated balance retrieval function to use fallback RPC clients as default instead of default wallet RPCs.
 - Rate conversion and display estimated token price for inputAsset.
 - Fixed fiat token prices.
+- Add Tokeo wallet support for SUI Mainnet
 
 ## [1.1.4]